30/8

+++ src/main/java/com/example/doan_tatruong/service/impl/OrderServiceImpl.java
@@ -16,6 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class OrderServiceImpl implements OrderService {
@@ -40,7 +41,7 @@ public class OrderServiceImpl implements OrderService {
         order.setUser(shoppingCart.getUser());
         order.setTotalPrice(shoppingCart.getTotalPrice());
         order.setAccept(false);
-        order.setPaymentMethod("Cash");
+        order.setPaymentMethod("Paid via PayPal");
         order.setOrderStatus("Pending");
         order.setQuantity(shoppingCart.getTotalItem());
         List<OrderDetail> orderDetailList = new ArrayList<>();
@@ -91,28 +92,49 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
+    @Transactional
     public void cancelOrder(Long id) {
-        orderRepository.deleteById(id);
-    }
+        Optional<Order> optionalOrder = orderRepository.findById(id);
 
+        if (optionalOrder.isPresent()) {
+            Order order = optionalOrder.get();
+
+            for (OrderDetail orderDetail : order.getOrderDetailList()) {
+                Product product = orderDetail.getProduct();
+                int orderedQuantity = orderDetail.getOrder().getQuantity();
+
+                product.setCurrentQuantity(product.getCurrentQuantity() + orderedQuantity);
+                productService.saveProduct(product);
+            }
+
+            orderRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Order with ID not found: " + id);
+        }
+    }
 
     @Override
     public void sendOrderConfirmationEmail(User user) {
         SimpleMailMessage message = new SimpleMailMessage();
         message.setTo(user.getUsername());
         message.setSubject("Order Confirmation");
-        message.setText("Thank you for your order!");
-
+        message.setText("Thank you for your order!" +
+                "link: http://localhost:8080/customer/order");
         javaMailSender.send(message);
     }
 
     @Override
     public Order updatePaymentMethod(Order order) {
-        if (order.getPaymentMethod().equals("PayPal")) {
+        if (order.getPaymentMethod().equals("paypal")) {
             order.setPaymentMethod("Paid via PayPal");
         } else {
             order.setPaymentMethod("Cash");
         }
         return orderRepository.save(order);
+    }
+
+    @Override
+    public void deleteOrder(Long id) {
+        orderRepository.deleteById(id);
     }
 }
